+++ PgsExt/PTData.cpp
@@ -1626,7 +1626,7 @@ HRESULT CPTData::Load(IStructuredLoad* pStrLoad,IProgress* pProgress)
    else
    {
       lrfdStrandPool* pPool = lrfdStrandPool::GetInstance();
-      Int32 key = var.lVal;
+      Int64 key = var.lVal;
       if ( version < 4 )
       {
          key |= matPsStrand::None; // add default encoding for stand coating type... added in version 4
@@ -1725,7 +1725,7 @@ HRESULT CPTData::Save(IStructuredSave* pStrSave,IProgress* pProgress)
    pStrSave->BeginUnit(_T("PTData"),4.0);
 
    lrfdStrandPool* pPool = lrfdStrandPool::GetInstance();
-   Int32 key = pPool->GetStrandKey(pStrand); // beginning with version 4, the strand key includes an including for strand coating type
+   Int64 key = pPool->GetStrandKey(pStrand); // beginning with version 4, the strand key includes an including for strand coating type
    pStrSave->put_Property(_T("TendonMaterialKey"),CComVariant(key));
 
    DuctIndexType ductCount = m_Ducts.size();
